start moving epics integration code into its own file
